Rename components according to CSS styles
