+++ ui/src/dashboards/components/CellEditorOverlay.js
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import uuid from 'node-uuid'
 
 import ResizeContainer, {ResizeBottom} from 'src/shared/components/ResizeContainer'
-import QueryBuilder from 'src/data_explorer/components/QueryBuilder'
+import QueryMaker from 'src/data_explorer/components/QueryMaker'
 import Visualization from 'src/data_explorer/components/Visualization'
 import OverlayControls from 'src/dashboards/components/OverlayControls'
 import * as queryModifiers from 'src/utils/queryTransitions'
@@ -159,7 +159,7 @@ class CellEditorOverlay extends Component {
                 onCancel={onCancel}
                 onSave={this.handleSaveCell}
               />
-              <QueryBuilder
+              <QueryMaker
                 source={source}
                 queries={queriesWorkingDraft}
                 actions={queryActions}
